feat: add subscriptionsMatcher Lambda (#559)
Also fixed a bunch of small things on the reaper lambda
