+++ package.json
@@ -43,6 +43,7 @@
     "fhir-works-on-aws-search-es": "3.9.2",
     "lodash": "^4.17.21",
     "serverless-http": "^2.7.0",
+    "tslib": "^2.3.1",
     "yargs": "^16.2.0"
   },
   "devDependencies": {
