+++ src/subscriptions/reaperLambda/index.ts
@@ -1,0 +1,6 @@
+import reaperHandler from './subscriptionReaper';
+
+/**
+ * Custom lambda handler that handles deleting expired subscriptions.
+ */
+exports.handler = reaperHandler;
