+++ src/subscriptions/reaperLambda/subscriptionReaper.ts
@@ -18,9 +18,13 @@ const reaperHandler = async (event: any) => {
     // filter out subscriptions without a defined end time.
     // check if subscription is past its end date (ISO format)
     // example format of subscriptions: https://www.hl7.org/fhir/subscription-example.json.html
+    console.log(subscriptions);
     return Promise.all(
         subscriptions
             .filter((s: Record<string, any>) => {
+                if (!s.end) {
+                    return false;
+                }
                 const date = new Date(s.end);
                 if (date.toString() === 'Invalid Date') {
                     console.log(`Skipping subscription ${s.id} since the end date is not in a valid format: ${s.end}`);
