+++ src/subscriptions/restHookLambda/index.ts
@@ -4,7 +4,6 @@
  *
  */
 
-import reaperHandler from './subscriptionReaper';
 import RestHookHandler from './restHook';
 import { AllowListInfo, getAllowListInfo } from './allowListUtil';
 
@@ -19,8 +18,3 @@ const restHookHandler = new RestHookHandler({ enableMultitenancy });
 exports.handler = async (event: any) => {
     return restHookHandler.sendRestHookNotification(event, allowListPromise);
 };
-
-/**
- * Custom lambda handler that handles deleting expired subscriptions.
- */
-exports.reaperHandler = reaperHandler;
